+++ Creditos/Models/mSobrante.cs
@@ -8,11 +8,11 @@ namespace Creditos.Models
     public class mSobrante{
         public int id_sobrante { get; set; }
         public int descuento_id { get; set; }
-        public float valor { get; set; }
+        public double valor { get; set; }
         public Boolean estado { get; set; }
 
         public mSobrante() { }
-        public mSobrante(int id_sobrante, int descuento_id, float valor, Boolean estado){
+        public mSobrante(int id_sobrante, int descuento_id, double valor, Boolean estado){
             this.id_sobrante = id_sobrante;
             this.descuento_id = descuento_id;
             this.valor = valor;
